Move IsControlFlowVisitor to file utils. test=develop
